==> notebooks/1-data-loading.ipynb <==
# **Dataset Loading**

# # **Objetivo desta Etapa**

Nesta etapa vamos carregar os dados originais e criar um arquivo pickle para facilitar futuras importações.

São 18 variáveis e 5.141.512 registros.

# # **Explicação do Dataset**

A base de dados foi obtida a partir do **Consumer Financial Protection Bureau (CFPB)**. Esses dados são referentes a reclamações de consumidores sobre produtos financeiros e serviços. O CFPB é uma agência independente do governo dos Estados Unidos responsável por proteger os consumidores no setor financeiro.

O site para obter o download é: https://www.consumerfinance.gov/data-research/consumer-complaints/

O arquivo baixado é um arquivo CSV (complaints.csv) que contém as seguintes colunas:

- Date received: data em que a reclamação foi recebida
- Product: produto financeiro ou serviço
- Sub-product: subproduto financeiro ou serviço
- Issue: problema relatado pelo consumidor
- Sub-issue: subproblema relatado pelo consumidor
- Consumer complaint narrative: descrição da reclamação
- Company public response: resposta pública da empresa
- Company: empresa envolvida na reclamação
- State: estado do consumidor
- ZIP code: código postal do consumidor
- Tags: tags relacionadas à reclamação
- Consumer consent provided: consentimento do consumidor fornecido
- Submitted via: forma de envio da reclamação
- Date sent to company: data em que a reclamação foi enviada para a empresa
- Company response to consumer: resposta da empresa ao consumidor
- Timely response: resposta oportuna
- Consumer disputed: consumidor contestado
- Complaint ID: ID da reclamação



---

# # **Carregando o Dataset**

import pandas as pd

# Isso carrega o CSV em pedaços menores e tenta inferir os tipos de dados com mais precisão. No entanto, pode consumir mais memória.
df = pd.read_csv('../dataset/original/complaints.csv', low_memory=False)

df.shape

# # **Criando Pickle**



df.to_pickle("./pickle/df_original.pkl")
