==> notebooks/6-data-final-processing.ipynb <==
# **Data Final Processing**

Nesta etapa iremos vamos preparar nosso dataset para ser utilizado no treinamento do modelo. 

---

# # **Importando Pickle da Etapa Anterior**

import pandas as pd

df = pd.read_pickle('./pickle/df_sales.pkl')

df.shape

df.columns

---

# # **Data Final Adjustments**

## ## **Criar Variável Chave para Modelo de Previsão de Indenização**

import pandas as pd

df['Company payed monetary relief'] = df['Company response to consumer'].apply(lambda x: True if x == 'Closed with monetary relief' else False)

df['Company payed monetary relief'].value_counts(), df.shape


## ## **Removendo Registros Nulos**

- Se em qualquer coluna o registro for nulo, o registro inteiro será removido.

# Contar o número de linhas antes de remover os nulos
linhas_antes = len(df)

# Remover as linhas que contêm valores nulos em qualquer coluna
df = df.dropna()

# Contar o número de linhas após remover os nulos
linhas_depois = len(df)

# Calcular o número de registros removidos
registros_removidos = linhas_antes - linhas_depois

# Exibir a quantidade de registros removidos
print(f"Número de registros removidos: {registros_removidos}")


---

df.to_excel('./excel/df_processed.xlsx', index=False)

df.to_pickle("./pickle/df_processed.pkl")

print("Notebook Final Processing Concluído")
