==> notebooks/8-combine-dataframes.ipynb <==
# **Combinando o DF original com o DF final**

import pandas as pd

df_original = pd.read_pickle('./pickle/df_cleaned.pkl')

df_final = pd.read_pickle('./pickle/df_predict_relief.pkl')



import pandas as pd

columns_to_add = ['Complaint ID', 'Complaint sentiment', 'Suggested Category', 'Suggested Product/Service','Suggested Reasoning','Potential Annual Revenue']

# Fazendo o merge dos DataFrames com base na coluna 'Complaints ID'
df_combined = pd.merge(df_original, df_final[columns_to_add], on='Complaint ID', how='left')

df_combined.shape

df_combined = df_combined.dropna(how='any')


df_combined.to_excel('./excel/df_combined.xlsx', index=False)
